Added CostCodeReport Feature
Added the Cost Code Reporting feature (includes several associated classes and enums) to the application.  This is in addition to the system-by-system comparison report feature.  In order to produce Grant's Labor Tracking report, a different set of parameters was needed for search control.  The key model class (CostCodeResult) contains the required information (as well as field for future extensibility).