Estimate Helper Complete
The Estimate Helper class is now functioning as intended; Medium-level validation has been included to accompany the formatting rules outlined in the Master Documentation.  The helper class attempts to calibrate the activeCell, then returns a list of SystemEstimate objects generated from the spreadsheet contents.

The ribbon code-behind consumes the list by sending it to the EstimateDataService class where it can be committed to the database.